Rename supersid source directory to src per git and github conventions
