+++ Config/supersid.cfg
@@ -35,7 +35,7 @@ waterfall_samples = 0
 
 hourly_save = YES
 
-# data_path shall be an absolute path or a path relative to the supersid script folder
+# data_path shall be an absolute path or a path relative to the src script folder
 data_path = ../Data
 
 paper_size = A4
@@ -87,7 +87,7 @@ Channels = 1
 automatic_upload = no
 ftp_server = sid-ftp.stanford.edu
 ftp_directory = /incoming/SuperSID/NEW/
-# local_tmp shall be an absolute path or a path relative to the supersid script folder
+# local_tmp shall be an absolute path or a path relative to the src script folder
 local_tmp = ../outgoing
 call_signs = NWC
