+++ src/supersid_icon.png
